Remove non-pauseless-threadstart code
This is more-or-less a rebased copy of 869f2ac, originally written by
<dougk@google.com>, which gets the pauseless-threadstart feature to work
on Windows.  Since Windows is the last outstanding platform that builds
with pauseless-threadstart disabled by default, this commit also removes
the pauseless-threadstart feature flag entirely.

The original commit was reverted by 15f2ccc due to failing tests.  The
only real difference between this commit and the original commit is that
this one sets th->no_tls_value_marker to NO_TLS_VALUE_MARKER in
alloc_thread_struct.