Add the :foreign-callback-fiber feature on Windows
:foreign-fiber-callback is an optional (:win32 + :sb-thread)-only
feature that uses the Windows fibers API to reduce the overhead of
foreign callbacks.

When the feature is enabled, foreign threads are automatically
partitioned into a foreign fiber and a callback fiber upon the first
foreign callback invocation. The foreign fiber executes the calling C
code, while the callback fiber runs foreign callbacks via a trampoline
loop running in a persistent Lisp thread. Suspending the callback
fiber after every call amortizes the cost of creating the Lisp thread
across many calls and reduces the steady-state cost of a foreign
callback to that of a call to SwitchToFiber.

This commit makes changes in several places to support this feature:
* Slots are added to the Lisp thread struct allowing the foreign fiber
  to pass the callback index, arguments pointer, and return pointer to
  the callback fiber. Slots are also added to hold the handles for the
  foreign and callback fibers.
* callback_wrapper_trampoline is modified to create the foreign and
  callback fibers on the first invocation from a non-Lisp thread, switch
  to the callback fiber in steady-state, and delete the fibers in the
  event that the callback trampoline loop dies.
* Functions for running the callback trampoline loop, switching to the
  foreign fiber, and entering a foreign callback from fiber resumption
  are added to alien-callback.lisp.

A slight modification is made to the fcb-threads tests so that they kill
the callback trampoline loop before completing. Also included in this
commit is a Windows-only benchmark measuring foreign callback overhead.