Add option to create PDB file for runtime on Win32
Microsoft debuggers such as WinDbg rely on a PDB (Program DataBase) file
to provide symbol information, source locations, and struct layouts for
a debuggee program. The debug information records within a PDB file are
encoded using another Microsoft format called CodeView. As of the time
of writing, only MSVC, Clang, and GCC >=14 can produce CodeView debug
information on Windows.

This commit adds a --with-runtime-pdb build option that does the
following:

* Use Clang instead of GCC.
* Produce CodeView debug info when compiling each runtime source file.
* Use -O3 since Clang doesn't have -O5.
* Link the CodeView debug info into a single sbcl.pdb file when linking
  sbcl.exe.
* Use __builtin_ffs because lld complains about "undefined symbol: ffs".